requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
openai>=1.3.7
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
streamlit>=1.28.0
scikit-learn
